Create lsShmQueue.php
This only works on linux, but it allows you to create a simple ring queue that I designed with the hopes of it working on Windows and Linux.  No dice on Windows, but the overall framework might come in handy to someone else. =)

The idea is to create a LsIpcMsgQueue object and then make a call into q->Receive().  There are two modes: polling and blocking.  To be honest, I haven't looked at this in over a year...  I know it 'worked' to a certain degree back in fall of 2014.  You might find this more useful as a collection of code fragements than a bone fide 'library'.  

Just sharing some work that may save some other unlucky bastard some time. =)