Create lsServer.php
This is a base class for a simple 'server' that processes incoming 'messages' and returns 'responses'.  Any server you write is just an extension of this class and needs to implement the methods:

    abstract protected function ServerSpecificProcessIncomingMessage( &$client );
    abstract protected function GetMainListeningPortNumber( );
    abstract protected function PeriodicTask( $serverUpSeconds );
    abstract protected function TimeSlice( );

Whenever a message is received, it's sent to ServerSpecificProcessIncomingMessage() and the LsServerClient object is sent in so the server subclass can issue a response on the same connection.

GetMainListeningPortNumber()
This method returns the port number that your derived LsServer subclass is listening on.

PeriodicTask()
This will get called approximately every second by the LsServer's main polling loop.

TimeSlice()
This gets caled everytime the LsServer's polling loop executes.  The interval of execution is 50ms by default:

LS_SERVER_POLL_USECS  = 50000;    /* 50ms polling period. */

You could certainly change this to whatever you like (or change this server model to an async mode, but I couldn't seem to get that working in php... no big deal for my application, so take from it what you will... =) )